requirements.txt
# python 3.8.13
numpy>=1.22.3
Cython>=0.29.28
cmake>=3.22.3    
flask==2.2.2
insightface
Pillow==9.2.0
# cpu
onnxruntime==1.12.0